Decode array before adding to changeset
Array values are encoded before added to the form for editing but
not decoded before evaluation by the changeset. This extra formatting
causes the changeset to mark arrays as invalid with no recourse by
the user. This commit modifies `Kaffy.ResourceSchema.get_map_fields/1`
to return `{:array, _}` fields to be decoded by
`Kaffy.ResourceParams.decode_map_fields/3`. Fixes #130 and #184