Add basic CI for supported Elixir versions (#221)
* Fix CI setup

This commit fixes the Github Actions config to use the
setup-beam actions package instead of  as
is suggested by Github.

* Update documentation to note Elixir 1.8 as minimum version

The matrix CI operation shows the use of the ceil/1. This method was added
to Elixir in 1.8 and effectively means that kaffy won't run on Elixir 1.7
as written. Since this is a relatively old version of Elixir, this updates
references to require Elixir 1.8 or higher.